requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy
passlib
bcrypt
sqlalchemy_utils
python-jose